feat: add `SearchRecordsQuery` schema
